Fixed issues with mempool cache
* A combination of the following items were causing rpc server to be
flooded/crash with larger blocks.

* Transaction processing is triggered by zmq and a cache is used when
 hash has already been seen.  When a block is found a new txhash
notification is received for each txn in the block, and the cache is
used to quickly ignore these.  Already seen non-SLP transactions in a
block are ignored using slpMempoolIgnoreList.  This cache was using
pop() instead of shift() for item removal which caused
SlpMempoolIgnoreList cache to not be update as required after
filling up to max size. Items would be added to the list with push()
and then removed immediately with pop(), instead of removing the
oldest item in the cache with shift().

* Recent commit added a new rpc call for getRawMempool on each zmq
notification for 'hashtx'.  This was removed since it is not needed since zmq
only includes mempool transactions.

* slpMempoolIgnore list is limited to 10000 items, with blocks larger
than 2MB this number may need to be larger.  When this limit is exceeded
every non-SLP zmq transaction notification (prior to block) will cause
a RPC call to getRawTransaction().  To fix this, we will want to switch to
using a raw transaction hex notification, rather than txn hash notification,
raw transaction hex notification.  Opened issue #34.

* slpMempoolIgnoreList was adding all transactions, not just non-SLP
transactions.